requirements.txt
opencv-python==4.8.1.78
torch==2.1.0
torchvision==0.16.0
Pillow==10.0.1
transformers==4.35.0
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
numpy==1.24.3
requests==2.31.0
llama-cpp-python==0.2.72 